requirements.txt
# Metadata System Requirements

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# File watching
watchdog>=3.0.0

# Data validation
pydantic>=2.0.0

# YAML parsing
pyyaml>=6.0

# LLM integration (already in main requirements)
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0